Use npm module for cron jobs rather than proper cron.
For scripts to run correctly we have to close DB connections at the
end of the process. This requires careful tracking of when every part
of the script is complete. If we can run cron inside of the app we
can resure that DB connection and thus simplify the code significantly.